build-src/requirements.txt
Flask
prometheus_client
python-bitcoinrpc
python-dotenv
